+++ python/agents/RAG/rag/agent.py
@@ -52,7 +52,7 @@ if rag_corpus:
             rag.RagResource(
                 # please fill in your own rag corpus
                 # here is a sample rag corpus for testing purpose
-                # e.g. projects/123/locations/us-east1/ragCorpora/456
+                # e.g. projects/123/locations/us-central1/ragCorpora/456
                 rag_corpus=rag_corpus
             )
         ],